backend/requirements.txt
# Django and Core
Django>=5.0,<5.1
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.3
drf-nested-routers>=0.93.4

# Database
psycopg2-binary>=2.9.0

# Real-time support
channels>=4.0.0
channels-redis>=4.1.0
redis>=5.0.0
websockets>=12.0  # For WebSocket testing

# Authentication & Authorization
djangorestframework-simplejwt>=5.3.0

# File uploads and media
Pillow>=10.0.0

# Environment variables
python-decouple>=3.8

# Development tools
django-debug-toolbar>=4.2.0

# API documentation
drf-spectacular>=0.26.0

# Utilities
django-extensions>=3.2.0

# Markdown support
markdown>=3.5.0

# Security and Rate Limiting
django-ratelimit>=4.0.0
django-csp>=3.7
